Issue #3461945 by mstrelan, bbrala: Fix instances of floats passed to functions expecting ints
(cherry picked from commit 4d805e5d9608cda37e8ef4282f82d42fba969907)

+++ core/lib/Drupal/Core/Cron.php
@@ -226,7 +226,7 @@ class Cron implements CronInterface {
       // Each queue will be processed immediately when it is reached for the
       // first time, as zero > currentTime will never be true.
       if ($process_from > $this->time->getCurrentMicroTime()) {
-        $this->usleep(round($process_from - $this->time->getCurrentMicroTime(), 3) * 1000000);
+        $this->usleep((int) round($process_from - $this->time->getCurrentMicroTime(), 3) * 1000000);
       }
 
       try {
